[Polyam] Account header split up
Port f2fb232e37d382308c4afa69fb189585306ec4c4 to polyam-glitch
